Implementing PIFCompiler
