+++ Sources/ScipioKit/PackageGraph+Extensions.swift
@@ -1,0 +1,8 @@
+import Foundation
+import PackageGraph
+
+extension ResolvedTarget {
+    var xcFrameworkName: String {
+        "\(c99name.packageNamed()).xcframework"
+    }
+}
